refactor: migrate from 'record' to 'recorder' package and consolidate related files
