+++ apps/chrome-extension/package.json
@@ -12,7 +12,7 @@
   "dependencies": {
     "@ant-design/icons": "^5.3.1",
     "@midscene/core": "workspace:*",
-    "@midscene/record": "workspace:*",
+    "@midscene/recorder": "workspace:*",
     "@midscene/report": "workspace:*",
     "@midscene/shared": "workspace:*",
     "@midscene/visualizer": "workspace:*",
