--- apps/chrome-extension/src/extension/record.tsx
@@ -1,2 +1,0 @@
-// Re-export the refactored Record component
-export { default } from './record/index';
