+++ apps/chrome-extension/src/extension/recorder/PLAYWRIGHT_GENERATOR.md
