+++ apps/chrome-extension/src/extension/recorder/README.md
