+++ apps/chrome-extension/src/extension/recorder/components/ProgressModal.tsx
