+++ apps/chrome-extension/src/extension/recorder/components/index.ts
