+++ apps/chrome-extension/src/extension/recorder/generators/README.md
