+++ apps/chrome-extension/src/extension/recorder/generators/index.ts
