+++ apps/chrome-extension/src/extension/recorder/generators/playwrightGenerator.ts
@@ -1,5 +1,5 @@
 import { generatePlaywrightTest as coreGeneratePlaywrightTest } from '@midscene/core/ai-model';
-import type { ChromeRecordedEvent } from '@midscene/record';
+import type { ChromeRecordedEvent } from '@midscene/recorder';
 import { recordLogger } from '../logger';
 import { handleTestGenerationError } from './shared/testGenerationUtils';
