+++ apps/chrome-extension/src/extension/recorder/generators/shared/testGenerationUtils.ts
@@ -1,4 +1,4 @@
-import type { ChromeRecordedEvent } from '@midscene/record';
+import type { ChromeRecordedEvent } from '@midscene/recorder';
 import { recordLogger } from '../../logger';
 import type {
   EventCounts,
