+++ apps/chrome-extension/src/extension/recorder/hooks/index.ts
