+++ apps/chrome-extension/src/extension/recorder/hooks/useRecordingSession.ts
