+++ apps/chrome-extension/src/extension/recorder/hooks/useTabMonitoring.ts
