+++ apps/chrome-extension/src/extension/recorder/logger.ts
