+++ apps/chrome-extension/src/extension/recorder/recorder.less
