+++ apps/chrome-extension/src/extension/recorder/shared/exportControlsUtils.ts
@@ -1,4 +1,4 @@
-import type { ChromeRecordedEvent } from '@midscene/record';
+import type { ChromeRecordedEvent } from '@midscene/recorder';
 import { message } from 'antd';
 import { useRecordStore, useRecordingSessionStore } from '../../../store';
 import { recordLogger } from '../logger';
