+++ apps/chrome-extension/src/extension/recorder/types.ts
