+++ apps/chrome-extension/src/extension/recorder/utils.ts
@@ -1,7 +1,7 @@
 import { AIActionType, callToGetJSONObject } from '@midscene/core/ai-model';
 import { message } from 'antd';
 
-import type { ChromeRecordedEvent } from '@midscene/record';
+import type { ChromeRecordedEvent } from '@midscene/recorder';
 import { recordLogger } from './logger';
 import { isChromeExtension, safeChromeAPI } from './types';
