+++ apps/record-form/package.json
@@ -13,7 +13,7 @@
     "dayjs": "^1.11.11",
     "react": "18.3.1",
     "react-dom": "18.3.1",
-    "@midscene/record": "workspace:*"
+    "@midscene/recorder": "workspace:*"
   },
   "devDependencies": {
     "@rsbuild/plugin-node-polyfill": "1.3.0",
