+++ packages/core/package.json
@@ -39,7 +39,7 @@
     "@azure/identity": "4.5.0",
     "@langchain/core": "0.3.26",
     "@midscene/shared": "workspace:*",
-    "@midscene/record": "workspace:*",
+    "@midscene/recorder": "workspace:*",
     "@ui-tars/action-parser": "1.2.0",
     "dirty-json": "0.9.2",
     "dotenv": "16.4.5",
