+++ packages/core/src/ai-model/prompt/yaml-generator.ts
@@ -1,4 +1,4 @@
-import type { ChromeRecordedEvent } from '@midscene/record';
+import type { ChromeRecordedEvent } from '@midscene/recorder';
 import {
   AIActionType,
   type ChatCompletionMessageParam,
