+++ packages/recorder/src/index.tsx
@@ -4,5 +4,5 @@ export {
   type RecordedEvent,
   type ChromeRecordedEvent,
   convertToChromeEvents,
-} from './record';
+} from './recorder';
 export { RecordTimeline } from './RecordTimeline';
